EvenOrOddAgent/pyproject.toml
[project]
name = "a2a-demo"
version = "0.1.0"
description = "Agent2Agent demo by hive agent"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.2.8",
    "uvicorn>=0.34.3"
]
